[Refactor] primary_action_sheet에 패딩 추가 및 show_farmus_bottom_sheet으로 이동 (#176)
